requirements.txt
pycryptodomex==3.10.1
pywin32==227
